Fixed Color Number range
Fixed the number range in the pixel processing to fit the 0...255
standard number range for colors

+++ Wrapper/RetroPortableUI.Wrapper.LibRetro/Support/PixelDefinition.cs
@@ -7,9 +7,9 @@ namespace RetroPortableUI.Wrapper.LibRetro.Support
 {
     public class PixelDefinition
     {
-        public int Alpha;
-        public int Red;
-        public int Green;
-        public int Blue;
+        public float Alpha;
+        public float Red;
+        public float Green;
+        public float Blue;
     }
 }
